Use gtest fixture for the newly added test
